backend/pyproject.toml
[tool.poetry]
name = "rttb-mongo"
version = "0.1.0"
description = ""
authors = ["jan-marais <janmarais@shoprite.co.za>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.16.0"
motor = "^2.5.1"
python-dotenv = "^0.19.2"
httpx = "^0.21.1"
black = "^21.12b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
